db: bump the connection pool size
The production worker is unable to run jobs:

ActiveRecord::ConnectionTimeoutError:
  could not obtain a connection from the pool within 5.000 seconds

This makes sense because the pool size is currently 5, and both Good
Job and Rails are being told to use 5 threads. Assuming the web server
is busy or keeps DB connections alive, it means the worker is starving
on the side.

Bump the pool size to the recommended calculation[1].

[1]: https://github.com/bensheldon/good_job/issues/669#issuecomment-1188022040